[CLS-1098] merging PathExpressionWithExcludeResolver to PathExpressionResolver
